refactor(caffeinate): separate inhibitors and settings
